+++ app/Http/Controllers/Api/UserController.php
@@ -6,19 +6,22 @@ use Carbon\Carbon;
 use App\Models\User;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
+use App\Http\Traits\responseApiTrait;
 use Illuminate\Support\Facades\Hash;
 
 class UserController extends Controller
 {
+    use responseApiTrait;
 
     public function __construct()
     {
-        $this->middleware('JWT');
+        $this->middleware('Admin');
     }
 
     public function index()
     {
-        return  User::all();
+        $user = User::get();
+        return $this->responseData('users', $user, 'Users Selected Successfully.');
     }
 
     public function store(Request $request)
@@ -26,7 +29,7 @@ class UserController extends Controller
 
         $request->validate([
             'name' => 'required',
-            'email' => ['required', 'email','unique:users'],
+            'email' => ['required', 'email', 'unique:users'],
             'password' => ['required', 'min:6'],
         ]);
 
@@ -37,53 +40,46 @@ class UserController extends Controller
             'created_at' => Carbon::now()->toDateTimeString(),
         ]);
 
-        // return $user;
-
-        return response()->json([
-            'user' => $user,
-            'status' => 201,
-            'message' => 'user created Successfully.'
-        ]);
+        return $this->responseSuccess('User Created Successfully.');
     }
 
     public function update(Request $request, $id)
     {
+        $user = User::find($id);
+
+        if (!$user) {
+            return $this->responseError('Not Found User', 404);
+        }
+
         $request->validate([
             'name' => 'required',
-            'email' => ['required', 'email','unique:users'],
+            'email' => ['required', 'email', 'unique:users'],
             'password' => ['sometimes', 'min:6'],
         ]);
 
         $id = User::findOrFail($id);
 
         $id->update($request->all());
-
-        return response()->json([
-            'status' => 200,
-            'message' => 'user Updated successfully.'
-        ]);
+        return $this->responseSuccess('User Updated Successfully.');
     }
 
     public function delete($id)
     {
 
-        $user = User::findOrFail($id);
+        $user = User::find($id);
+
+        if (!$user) {
+            return $this->responseError('Not Found User', 404);
+        }
 
         $image = $user->profile->image;
 
-        if($image){
-
+        if ($image) {
             unlink($image);
-
         }
         if ($user) {
-
             $user->delete();
-            return response()->json([
-                'status' => 200,
-                'message' => 'User Deleted Successfully.'
-            ]);
+            return $this->responseSuccess('User Deleted Successfully.');
         }
-
     }
 }
